refactor(daemon): extract run-telemetry + chat-request composition from server.ts
Strangler-fig slice 3: split the finalized-run telemetry and turn-2
form-answer/chat-request composition helpers out of server.ts into two cohesive
sibling modules, byte-identical (sliced from source to preserve the — / CJK
escape sequences and em-dash comments exactly).

- run-telemetry.ts: shouldReportRunCompletedFromMessage, telemetryPromptFromRunRequest,
  createFinalizedMessageTelemetryReporter (dependency-injected design/db/dataDir/
  reportedRuns), shouldReportRunCompletionTelemetryFallbackStatus, and the
  TERMINAL_RUN_STATUSES set (moved here as its sole consumer; private).
- chat-request-composition.ts: FORM_ANSWERS_HEADER_RE, FORM_ANSWERED_SYSTEM_OVERRIDE,
  FORM_ANSWERED_GENERIC_OVERRIDE, formAnswerTransitionForCurrentPrompt,
  composeChatUserRequestForAgent (all pure).

server.ts imports back the symbols it references (incl. FORM_ANSWERS_HEADER_RE for
the chat-run deps object) and re-exports the public surface. Deps are all imports
(langfuse-bridge, run-result, contracts/analytics) -- no daemon module state.

No behavior change; chat-route form-answer wording asserted green. Orphan/duplicate
checks clean.
